refactor: bring load_costs features into build_cost_data
