+++ config/config.default.yaml
@@ -894,19 +894,6 @@ costs:
     "discount rate": 0.07
   custom_costs: false
   overwrites: {}
-  marginal_cost:
-    solar: 0.01
-    onwind: 0.015
-    offwind: 0.015
-    hydro: 0.
-    H2: 0.
-    electrolysis: 0.
-    fuel cell: 0.
-    battery: 0.
-    battery inverter: 0.
-    home battery storage: 0
-    water tank charger: 0.03
-    central water pit charger: 0.025
   emission_prices:
     enable: false
     co2: 0.
