+++ rules/build_sector.smk
@@ -1273,7 +1273,6 @@ rule prepare_sector_network:
             "existing_capacities", "conventional_carriers"
         ),
         foresight=config_provider("foresight"),
-        costs=config_provider("costs"),
         sector=config_provider("sector"),
         industry=config_provider("industry"),
         renewable=config_provider("renewable"),
@@ -1346,9 +1345,11 @@ rule prepare_sector_network:
             "biomass_potentials_s_{clusters}_{planning_horizons}.csv"
         ),
         costs=lambda w: (
-            resources("costs_{}.csv".format(config_provider("costs", "year")(w)))
+            resources(
+                "costs_{}_prepped.csv".format(config_provider("costs", "year")(w))
+            )
             if config_provider("foresight")(w) == "overnight"
-            else resources("costs_{planning_horizons}.csv")
+            else resources("costs_{planning_horizons}_prepped.csv")
         ),
         h2_cavern=resources("salt_cavern_potentials_s_{clusters}.csv"),
         busmap_s=resources("busmap_base_s.csv"),
