+++ rules/collect.smk
@@ -15,10 +15,10 @@ localrules:
 rule build_costs:
     input:
         lambda w: (
-            [resources(f"costs_{config_provider('costs', 'year')(w)}_extended.csv")]
+            [resources(f"costs_{config_provider('costs', 'year')(w)}_prepped.csv")]
             if config_provider("foresight")(w) == "overnight"
             else expand(
-                resources("costs_{planning_horizons}_extended.csv"),
+                resources("costs_{planning_horizons}_prepped.csv"),
                 **config["scenario"],
                 run=config["run"]["name"],
             )
