+++ rules/solve_myopic.smk
@@ -21,7 +21,7 @@ rule add_existing_baseyear:
         busmap=resources("busmap_base_s_{clusters}.csv"),
         clustered_pop_layout=resources("pop_layout_base_s_{clusters}.csv"),
         costs=lambda w: resources(
-            "costs_{}.csv".format(
+            "costs_{}_prepped.csv".format(
                 config_provider("scenario", "planning_horizons", 0)(w)
             )
         ),
@@ -87,7 +87,7 @@ rule add_brownfield:
             "networks/base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}.nc"
         ),
         network_p=solved_previous_horizon,  #solved network at previous time step
-        costs=resources("costs_{planning_horizons}.csv"),
+        costs=resources("costs_{planning_horizons}_prepped.csv"),
         cop_profiles=resources("cop_profiles_base_s_{clusters}_{planning_horizons}.nc"),
     output:
         resources(
@@ -125,7 +125,7 @@ rule solve_sector_network_myopic:
         network=resources(
             "networks/base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}_brownfield.nc"
         ),
-        costs=resources("costs_{planning_horizons}.csv"),
+        costs=resources("costs_{planning_horizons}_prepped.csv"),
     output:
         network=RESULTS
         + "networks/base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}.nc",
