+++ scripts/_helpers.py
@@ -1075,3 +1075,21 @@ def load_cutout(
         cutout.data = cutout.data.sel(time=time)
 
     return cutout
+
+
+def load_costs(cost_file: str) -> pd.DataFrame:
+    """
+    Load cost data from CSV and prepare it.
+
+    Parameters
+    ----------
+    cost_file : str
+        Path to the CSV file containing cost data
+
+    Returns
+    -------
+    costs : pd.DataFrame
+        DataFrame containing the processed cost data
+    """
+
+    return pd.read_csv(cost_file, index_col=0).sort_index()
