+++ scripts/build_cost_data.py
@@ -2,7 +2,7 @@
 #
 # SPDX-License-Identifier: MIT
 """
-Extends cost data with custom cost modifications.
+Prepare and extend cost data with custom cost modifications.
 
 Inputs
 ------
@@ -13,10 +13,17 @@ Inputs
 Outputs
 -------
 
-- ``resources/costs_{planning_horizons}_extended.csv``: Extended cost data with custom modifications applied
+- ``resources/costs_{planning_horizons}_prepped.csv``: Prepared cost data with custom modifications applied
 """
 
+import logging
+
 import pandas as pd
+from _helpers import get_snapshots
+
+from scripts.add_electricity import calculate_annuity
+
+logger = logging.getLogger(__name__)
 
 
 def expand_all_technologies(
@@ -65,11 +72,122 @@ def expand_all_technologies(
     return custom_costs_expanded
 
 
+def prepare_costs(
+    costs: pd.DataFrame, config: dict, max_hours: dict = None, nyears: float = 1.0
+) -> pd.DataFrame:
+    """
+    Prepare costs to standardize input costs data.
+
+    Parameters
+    ----------
+    costs
+    config : dict
+        Dictionary containing cost-related configuration parameters
+    max_hours : dict, optional
+        Dictionary specifying maximum hours for storage technologies
+    nyears : float, optional
+        Number of years for investment, by default 1.0
+
+    Returns
+    -------
+    costs : pd.DataFrame
+        DataFrame containing the prepared cost data
+
+    """
+    # Copy marginal_cost and capital_cost for backward compatibility
+    for key in ("marginal_cost", "capital_cost"):
+        if key in config:
+            config["overwrites"][key] = config[key]
+
+    # set all asset costs and other parameters
+    costs = costs.set_index(["technology", "parameter"]).sort_index()
+
+    # correct units to MW and EUR
+    costs.loc[costs.unit.str.contains("/kW"), "value"] *= 1e3
+    costs.loc[costs.unit.str.contains("/GW"), "value"] /= 1e3
+
+    costs.unit = costs.unit.str.replace("/kW", "/MW")
+    costs.unit = costs.unit.str.replace("/GW", "/MW")
+
+    # min_count=1 is important to generate NaNs which are then filled by fillna
+    costs = costs.value.unstack(level=1).groupby("technology").sum(min_count=1)
+    costs = costs.fillna(config["fill_values"])
+
+    # Process overwrites for various attributes
+    for attr in ("investment", "lifetime", "FOM", "VOM", "efficiency", "fuel"):
+        overwrites = config["overwrites"].get(attr)
+        if overwrites is not None:
+            overwrites = pd.Series(overwrites)
+            costs.loc[overwrites.index, attr] = overwrites
+            logger.info(f"Overwriting {attr} with:\n{overwrites}")
+
+    annuity_factor = calculate_annuity(costs["lifetime"], costs["discount rate"])
+    annuity_factor_fom = annuity_factor + costs["FOM"] / 100.0
+    costs["capital_cost"] = annuity_factor_fom * costs["investment"] * nyears
+
+    costs.at["OCGT", "fuel"] = costs.at["gas", "fuel"]
+    costs.at["CCGT", "fuel"] = costs.at["gas", "fuel"]
+
+    costs["marginal_cost"] = costs["VOM"] + costs["fuel"] / costs["efficiency"]
+
+    costs.at["OCGT", "CO2 intensity"] = costs.at["gas", "CO2 intensity"]
+    costs.at["CCGT", "CO2 intensity"] = costs.at["gas", "CO2 intensity"]
+
+    costs.at["solar", "capital_cost"] = costs.at["solar-utility", "capital_cost"]
+    costs = costs.rename({"solar-utility single-axis tracking": "solar-hsat"})
+
+    # Calculate storage costs if max_hours is provided
+    if max_hours is not None:
+
+        def costs_for_storage(store, link1, link2=None, max_hours=1.0):
+            capital_cost = link1["capital_cost"] + max_hours * store["capital_cost"]
+            if link2 is not None:
+                capital_cost += link2["capital_cost"]
+            return pd.Series(
+                {
+                    "capital_cost": capital_cost,
+                    "marginal_cost": 0.0,
+                    "CO2 intensity": 0.0,
+                }
+            )
+
+        costs.loc["battery"] = costs_for_storage(
+            costs.loc["battery storage"],
+            costs.loc["battery inverter"],
+            max_hours=max_hours["battery"],
+        )
+        costs.loc["H2"] = costs_for_storage(
+            costs.loc["hydrogen storage underground"],
+            costs.loc["fuel cell"],
+            costs.loc["electrolysis"],
+            max_hours=max_hours["H2"],
+        )
+
+    for attr in ("marginal_cost", "capital_cost"):
+        overwrites = config["overwrites"].get(attr)
+        if overwrites is not None:
+            overwrites = pd.Series(overwrites)
+            idx = overwrites.index.intersection(costs.index)
+            costs.loc[idx, attr] = overwrites.loc[idx]
+            logger.info(f"Overwriting {attr} with:\n{overwrites}")
+
+    return costs
+
+
 if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        snakemake = mock_snakemake("build_cost_data", planning_horizons=2030)
+        snakemake = mock_snakemake(
+            "build_cost_data", clusters=5, planning_horizons=2030
+        )
+
+    config = snakemake.params.costs
+
+    snapshots = get_snapshots(
+        snakemake.params.snapshots, snakemake.params.drop_leap_day, tz="UTC"
+    )
+    nyears = len(snapshots) / 8760
 
     # Retrieve costs assumptions
     planning_horizon = str(snakemake.wildcards.planning_horizons)
@@ -80,7 +198,7 @@ if __name__ == "__main__":
         .drop("planning_horizon", axis=1)
     )
 
-    if snakemake.params.custom_costs and not custom_costs.empty:
+    if config.get("custom_costs", False) and not custom_costs.empty:
         # Expand "all" technologies across all available technologies from default costs
         custom_costs_expanded = expand_all_technologies(costs, custom_costs)
 
@@ -91,4 +209,9 @@ if __name__ == "__main__":
     else:
         costs_extended = costs
 
-    costs_extended.to_csv(snakemake.output[0], index=False)
+    # Prepare costs
+    costs_prepped = prepare_costs(
+        costs_extended, config, snakemake.params.max_hours, nyears
+    )
+
+    costs_prepped.to_csv(snakemake.output[0])
